% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_funs.R
\name{chol_inv}
\alias{chol_inv}
\title{chol_inv - a function to compute inverse of a symmetric matrxi}
\usage{
chol_inv(x)
}
\value{
A matrix
}
\description{
chol_inv - a function to compute inverse of a symmetric matrxi
}
\details{
More summary needed here.
}
