\name{DNC-package}
\alias{DNC-package}
\alias{DNC}
\docType{package}
\title{
\packageTitle{DNC}
}
\description{
\packageDescription{DNC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DNC}
\packageIndices{DNC}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{DNC}

Maintainer: \packageMaintainer{DNC}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
