\name{friendship}
\alias{friendship}
\docType{data}
\title{
Longitudinal classroom friendship network (Andrea Knecht)
}
\description{
This friendship network consists of 26 students in a middle school in the Netherlands measured at four time points throughout a school year.
}
\usage{data("friendship")}
\format{
The data is a 26 x 26 x 4 array of 1's and 0's (and NA's).
}
\details{
See \code{knecht {xergm.common}} for more details, as well as for behavioral and demographic actor attributes.
}
\references{
Knecht, Andrea (2006): Networks and Actor Attributes in Early Adolescence [2003/04]. Utrecht, The Netherlands Research School ICS, Department of Sociology, Utrecht University. (ICS-Codebook no. 61).

Knecht, Andrea (2008): Friendship Selection and Friends' Influence. Dynamics of Networks and Actor Attributes in Early Adolescence. PhD Dissertation, University of Utrecht. http://dspace.library.uu.nl/bitstream/handle/1874/25950/full.pdf.
}
\examples{
data(friendship)
}
\keyword{datasets}
