\name{Internals}
\alias{cLogLikelihood}
\alias{cVBUpdate}
\alias{cVBnoClust}
\alias{cllikProj}
\alias{cposteriorNoOmega}
\alias{csimulateData}
\alias{drawMiss}
\alias{drawS}
\alias{drawTau}
\alias{drawU}
\alias{drawXZar}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal dnc functions
}
\description{
Internal functions called by \code{dnc}
}

\details{
These functions are c++ not intended to be called by the user.
}
