% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_funs.R
\name{log_post_fun}
\alias{log_post_fun}
\title{log_post_fun - a function to compute the log posterior}
\usage{
log_post_fun(params, x, y, mu, sigma)
}
\value{
A scalar
}
\description{
log_post_fun - a function to compute the log posterior
}
\details{
More summary needed here.
}
