% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_funs.R
\name{log_post_fun_dnc}
\alias{log_post_fun_dnc}
\title{log_post_fun_dnc - a function to compute the log posterior using power likelihood}
\usage{
log_post_fun_dnc(params, x, y, mu, sigma, ratio)
}
\value{
A scalar
}
\description{
log_post_fun_dnc - a function to compute the log posterior using power likelihood
}
\details{
More summary needed here.
}
