% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_funs.R
\name{MH}
\alias{MH}
\title{MH - perform MH alogrithm}
\usage{
MH(MH_draws, proposal_cov, x, y, mu, sigma, ratio)
}
\value{
A integer of the number of MH samples accepted
}
\description{
MH - perform MH alogrithm
}
\details{
Used log_post_fun_dnc so if you don't want DNC set ratio = 1.
Modifies supplied MH_draws matrix in place.
}
